2021 Day 20
Part 1 and 2

Aaaa! I started with a set of dots, changed to a dict (with true/false value) so that I could detect coords out-of-bounds, but I didn't change the count at the end! I had the correct solution for so long, bu I counted all squares instead of just the active ones! Aaaaaaaa!