2016 Day 11
Part 1 and 2
6 hours for part 1 and 20 minutes for part 2. After getting the example
to work with hash sets I let it run for half an hour, then made some
(tiny) improvements and let it run for an hour while eating. After that
I rewrote it as bit-fields and now it finishes part 1 in 2 seconds, and
part 2 in one minute.
I kept the example elements in the code for part 1, and that screwed me
over a bit on part 2. With 10 elements the bit-filed required 72 bits,
and I didn't realise that immediately.