2016 Day 16
Part 1 and 2.
Several days. Part 1 solution took 20 minutes max. I gave up on part 2
after having run the part 1 solution on it for about two hours. When I
came back to it, it took maybe and hour of unfocused work. But I've been
thinking about it on and off for a few days.