2022 Day 8
Part 1 and 2.
I did some much wrong today.

I had `height = 0` as start in part 1, wrote some visualizing code for
the example, saw the `0` and fixed it in one place (!), had a bug in the
visualizer and took forever to realize that the `0` were in more places.

Then for part 2, I was stuck on calculating too many threes. I read in
the example that the view was blocked "by a massive tree of height `9`",
and thought that "the world is surrounded by height `9` trees".
Something which is complete fiction, I made it up. So I was stuck on
that for a while before realizing that, no trees outside doesn't count,
and edge trees will always have zero in one direction.

Not happy with today. Not as fun problem as yesterday, and no good
placing. Four thousand something. And I even got bronze on Cygnis
leaderboard.