2022 Day 9
Part 1 and 2.
Was stuck for 15 min extra on part 1 because I decided to not include a
`+` in the regex. Neither the example or the first page of real input
had more than one digit in the distance, so why include one?
Also had some complications with my `Follow` function for part 2 but I
can't blame that on any thing.