2022 Day 15
Part 1 and 2.
I tried to use my `NeighborsVonNeumann` with an insanely large range
for both parts but it was too slow. Forgot the `Math.Abs` for part 1 and
forgot to offset the point by the sensor for part 2. Could have gone
faster if I didn't mess up, but I'm still pretty happy with today.