2022 Day 20
Part 1 and 2.
Aaaa! I did solution 3&4 first, but I removed the node from the list
after iterating over it to find it's new place. In the example we never
moved "over" the current node so I didn't notice it! Redid it with
indices and an array list and probably did the same thing wrong, before
I did solution 1&2, where I worked after I put in `Length - 1`. Didn't
understand why, but 🤷 When I redid 3&4 I changed it to only move
forward (and modding the value by the list count), and that's when I
questioned if the current node should be included in the list count.
Guess you learn something every day 🙃