2022 Day 24
Part 1 and 2.
I forgot to add the bounds check in `MakeMoves`, found it while making a
print function for debugging but continued with the printing. When I was
done, I thought the example were supposed to evaluate to 14 and
continued looking for the problem even though I had the correct
solution. Could have been a bit faster...
i'm also like "This isn't difficult at all", I know I have my path
finding library, but this feels like day 10-15 level. After that extream
weekend all the problems have been easy., and tomorrow's not supposed to
be difficult so, I guess that's it 🤷