2023 Day 17
Almost on the board for part 1! I forgot in which order my library took
it's arguments and provided the "cost" in the "dierction" spot. And then
I messed somthing up with the steps count. Why is it that for part 1 it
works with `steps < 4` but for part 2 it works with `steps <= 10`?
Weird...