2024 Day 11 Part 2
I read a suggestion from Kalle. This was super smart!

I didn't think this would be fast so I didn't think in this direction. I
was stuck on doing some kind of memoization or something, mixing in the
stone number combined with remaining blinks. But I don't know. I
couldn't figure that out.